feedbackModel/RecommendationModel/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
transformers==4.35.2
torch==2.1.1
pinecone-client
httpx
supabase==1.0.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
python-dotenv==1.0.0

# PostgreSQL
psycopg2-binary==2.9.9
